# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's Moon Group Transparency",
        "description": "API's for transparency of Moon Group",
        "version": "1.0.0"
    },
    "paths": {
...
        "/moontransparency/public/api/survey": {
            "get": {
                "tags": [
                    "Survey"
                ],
                "summary": "Obtener información de Surveys con filtros y ordenamiento",
                "operationId": "1c581aae899aaad5f2562acfcf56c248",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Fecha de inicio",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Fecha de fin",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "ID del proyecto",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "survey_name",
                        "in": "query",
                        "description": "Nombre de la encuesta",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Descripción de la encuesta",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de Surveys",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Survey"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Survey"
                ],
                "summary": "Crear Survey",
                "operationId": "43edc69bc9f3ac8a24ac5c86f0c7d52e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/SurveyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Donación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Survey"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "La validación falló."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/moontransparency/public/api/survey/{id}": {
            "get": {
                "tags": [
                    "Survey"
                ],
                "summary": "Obtener detalles de un Survey por ID",
                "operationId": "b0eb0d41be562b1fba1997326d2544f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del Survey",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donación encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Survey"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donación no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Donación no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Survey"
                ],
                "summary": "Actualizar un Survey",
                "operationId": "3c76011836495748da8ccea6b9ac0b7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/SurveyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Donación actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Survey"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Datos inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donación no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Donación no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Survey"
                ],
                "summary": "Eliminar un Survey por ID",
                "operationId": "fee51a3d8ac48a1f6e2b69e349fbc549",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donación eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Donación eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Donación no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "SurveyRequest": {
                "required": [
                    "project_id",
                    "survey_name",
                    "description"
                ],
                "properties": {
                    "project_id": {
                        "type": "integer",
                        "example": "101"
                    },
                    "survey_name": {
                        "type": "string",
                        "example": "Encuesta de Energía Renovable"
                    },
                    "description": {
                        "type": "string",
                        "example": "Encuesta para evaluar el uso de energía renovable en zonas rurales."
                    }
                },
                "type": "object"
            },
            "Survey": {
                "title": "Survey",
                "description": "Modelo de Survey",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "project_id": {
                        "type": "integer",
                        "example": "101"
                    },
                    "survey_name": {
                        "type": "string",
                        "example": "Encuesta de Energía Renovable"
                    },
                    "description": {
                        "type": "string",
                        "example": "Encuesta para evaluar el uso de energía renovable en zonas rurales."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-26T21:44:24"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}